+++ GymTrain/Models/Result.cs
@@ -10,17 +10,17 @@ namespace GymTrain.Models
         public DateTime Date { get; set; }
         public int RestingHeartRate { get; set; }
         public int Height { get; set; }
-        public int Weight { get; set; }
-        public int BMI { get; set; }
-        public int Neck { get; set; }
-        public int Bust { get; set; }
-        public int Chest { get; set; }
-        public int UpperArm { get; set; }
-        public int Forearm { get; set; }
-        public int Waist { get; set; }
-        public int Hips { get; set; }
-        public int Thigh { get; set; }
-        public int Calves { get; set; }
+        public decimal Weight { get; set; }
+        public decimal BMI { get; set; }
+        public decimal Neck { get; set; }
+        public decimal Bust { get; set; }
+        public decimal Chest { get; set; }
+        public decimal UpperArm { get; set; }
+        public decimal Forearm { get; set; }
+        public decimal Waist { get; set; }
+        public decimal Hips { get; set; }
+        public decimal Thigh { get; set; }
+        public decimal Calves { get; set; }
 
         public ICollection<Result> Results { get; set; }
     }
